Reclamo de recompensas, vista de historial de todas las recompensas recibidas

+++ lib/MediaQuery.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class Pantalla{
 
   ///(Devuelve un porcentaje del tamaño de la pantalla. (porcentaje, contexto, eje(y: largo, x:ancho))
-  static double getPorcentPanntalla(int porcentaje, BuildContext context, String eje){
+  static double getPorcentPanntalla(double porcentaje, BuildContext context, String eje){
 
     if(eje == 'y'){
       return (porcentaje/100) * (MediaQuery.of(context).size.height);
